requirements.txt
langchain>=0.3
langchain-openai>=0.2
langgraph>=0.2
chainlit>=1.2
pydantic>=2
python-dotenv>=1
